Group pytest's tests (elementary, files).
The introduction of a pytest.ini appears as a convienient approach
to group tests of pytests sharing an element in common.  This seems
beneficial as long as there is no reliable approch identified to
stack "file" and "folder" as levels of an additional parameter in
theses tests for date2name.

By now, the two layers "elementary" and "files" are installed.