redefine month pattern (explicit day truncation)
Regexes' grouping with parentheses and the definition of the
standard pattern just ahead of the month pattern allowed to spot
a plausible cause for the problem to retract all autogenerated
stamps.  In the present scope of pytest, the additional lookafter
now sufficies to enable the retraction of all five stamp pattern.